Fetch all recording function is now implemented in MVVM Architecture.
