+++ FullCast/Model/AlertDetails.swift
@@ -1,0 +1,13 @@
+//
+//  AlertDetails.swift
+//  FullCast
+//
+//  Created by Vishwa  R on 07/02/22.
+//
+
+import Foundation
+
+struct AlertDetails {
+    var alertTitle : String
+    var alertMessage : String
+}
